fixes, progress

+++ rpython/translator/revdb/src-revdb/revdb_preinclude.h
@@ -1,8 +1,8 @@
 #include <stdint.h>
 
 typedef struct rpy_revdb_command_s {
-    int cmd;      /* neg for standard commands, pos for interp-specific */
-    size_t extra_size;
+    int32_t cmd;      /* neg for standard commands, pos for interp-specific */
+    uint32_t extra_size;
     int64_t arg1;
     int64_t arg2;
     int64_t arg3;
